move common/common to root common
